+++ sidplay-residfp/libsidplayfp/src/utils/STILview/stil.cpp
@@ -669,7 +669,7 @@ STIL::getDirs(ifstream &inFile, dirList &dirs, bool isSTILFile)
             if (strncmp(line.data(), "#  STIL v", 9) == 0)
             {
                 // Get the version number
-                STILVersion = atof(line.c_str() + 9);
+                STILVersion = (float)atof( line.c_str() + 9 );
 
                 // Put it into the string, too.
                 ostringstream ss;
